Implement zookeeper LWRP.
Switch the Zookeeper default recipe to an LWRP. Maintain backwards
compatibility by retaining the default recipe and calling the zookeeper
resource from within it.

The LWRP provides the capacity to alter user, version, mirror, checksum,
install directory, and to autostart the service when done.

This converges on Ubuntu 12.04. Testing for 14.04 pending. CentOS did
not work, but this is still slightly WIP.

Also rearranged attributes and bumped default ZK version to 3.4.6.