Add default attribute for config and render by default.
This adds a client port, tick time, and data directory reflective of the
official getting started guide for Zookeeper. It also renders the config
in the default recipe, such that the only missing piece now is a service
hook (which will not be included in the recipe but be available as a
resource).