Add install-only recipe[v3]
Added functionality to only create zookeeper user and install files with
a new recipe 'install' which is just a subset of the default recipe.
The default recipe continues to work as it did before.

This is a respin of the previous patch with a couple of changes to make
merging easier for upstream.

Signed-off-by: Scott Lovenberg <scott.lovenberg@gmail.com>