Always include build-essential
Build tools including GCC, patch, etc must be installed in order to
build the zookeeper and json gems.  `build-essential` handles this in
the `install` recipe, however the `build-essential` recipe is *only*
included when the `use_java_cookbook` attribute is set to `true`.

This patch unconditionally includes the `build-essential` recipe.

This was previously fixed in the `exhibitor` recipe for issue #73/#74,
however the exhibitor recipe no longer exists.

Tested with CentOS 6.6 and Ubuntu 14.04.