Fix up Test Kitchen syntax
The previous syntax would set Ubuntu 12.04 to test with the specified
Chef version, leaving the two CentOS platforms to use the latest version of
Chef (since the default value of `require_chef_omnibus` is `true`).

This now specifies that we use the latest version of Chef 11.8 for all tests,
providing more consistency.