Use prerelease runit to fix race condition
It seems that, in the current latest release of runit, there is a bug
that causes the service enable/start to fail on first convergence:

sv: unable to change to service directory: file does not exist

This makes testing tricky. For now, we’ll use this ref of the cookbook,
and when it finally ships, we can drop this and resume using the latest
Supermarket release.