Change behavior of env_vars & its guard
Thanks to the ChefSpec tests, this method was revealed to not be valid,
since the compilation phase would try to run a value of `false` through
the `exports_config` method, which would not work.

This results in the same output, with a valid default value to pass to
`exports_config` during compilation.