Use java cookbook’s version of Java
The java cookbook sets $JAVA_HOME to the absolute path for the version
of Java it installs, and this should be used to ensure we are, in fact,
running with the correct installation of Java.

This combats a situation where one might install multiple JDKs on a
machine (or multiple JDKs are installed as dependencies by other
software packages).